Split functional tests into different files
